Move calc_hash
