+++ spack/src/lib.rs
@@ -6,4 +6,5 @@ extern crate test;
 mod bundler;
 pub mod config;
 mod debug;
+mod hash;
 pub mod loaders;
